Add dependencies for GigRouter compatibility
This changeset introduces packages which are dependencies for Cambrian Works GigRouter applications.
Package lists provided as a common include to Kas configurations. Microcom and gnupg patch recipes
are added under recipes-support, since the former isn't included as part of OE. A collection of
recipes-dummy is added to mimic packages which CW applications have dependencies on but the content
of which is provided through other packages in meta-tegra (e.g., cuda-toolkit-11-4 is superseded by
cuda-toolkit, etc).