Add GigCompute layer
This changeset splits the meta-gigrouter layer into three new layers:

1) A meta-gig layer which is common to all GigXXX products, providing
common hardware configuration and software components.

2) A meta-gigrouter layer which contains GigRouter-specific recipes such
as user setup and, in future, hardware diversifications.

3) A meta-gigcompute layer which contains GigCompute-specific recipes such
as user setup and, in future, hardware diversification.

The top-level build script is updated to support intiating a GigCompute
build in addition to the existing GigRouter targets