Set hostname automcatically
This chanegset introduces a script, called through systemd prior to
network bringup, which sets the units's hostname to a unique name derived
from the SOM's unique serial ID if it hasn't already been set. Format is:

gr-XXXX for GigRouter images.
gc-XXXX for GigCompute images.

The script is only executed one time on first boot. This prevents overwriting
the hostname if a customer has manually set it later.