Create nvidia-l4t-nvpmodel dummy package (#45)
* Create nvidia-l4t-nvpmodel dummy package

nvidia-l4t-nvpmodel is a package provided by the Ubuntu SDK which is
named differently in Yocto. This changeset introduces a dummy package
to satisfy other package's dependencies on the old naming.

This changeset also extends build coverage for GigCompute builds to
Orin Nano boards.

* Create nvidia-l4t-nvpmodel dummy package

nvidia-l4t-nvpmodel is a package provided by the Ubuntu SDK which is
named differently in Yocto. This changeset introduces a dummy package
to satisfy other package's dependencies on the old naming.

This changeset also extends build coverage for GigCompute builds to
Orin Nano boards.