Set hostname automatically (#46)
* Set hostname automcatically

This chanegset introduces a script, called through systemd prior to
network bringup, which sets the units's hostname to a unique name derived
from the SOM's unique serial ID if it hasn't already been set. Format is:

gr-XXXX for GigRouter images.
gc-XXXX for GigCompute images.

* Create nvidia-l4t-nvpmodel dummy package (#45)

* Create nvidia-l4t-nvpmodel dummy package

nvidia-l4t-nvpmodel is a package provided by the Ubuntu SDK which is
named differently in Yocto. This changeset introduces a dummy package
to satisfy other package's dependencies on the old naming.

This changeset also extends build coverage for GigCompute builds to
Orin Nano boards.

* Create nvidia-l4t-nvpmodel dummy package

nvidia-l4t-nvpmodel is a package provided by the Ubuntu SDK which is
named differently in Yocto. This changeset introduces a dummy package
to satisfy other package's dependencies on the old naming.

This changeset also extends build coverage for GigCompute builds to
Orin Nano boards.

* Set hostname automcatically

This chanegset introduces a script, called through systemd prior to
network bringup, which sets the units's hostname to a unique name derived
from the SOM's unique serial ID if it hasn't already been set. Format is:

gr-XXXX for GigRouter images.
gc-XXXX for GigCompute images.

The script is only executed one time on first boot. This prevents overwriting
the hostname if a customer has manually set it later.